+++ PhD_RAG/src/database/services.py
@@ -1,3 +1,4 @@
+from langchain_text_splitters import MarkdownHeaderTextSplitter
 from langchain_milvus import Milvus
 from PhD_RAG.src.database.embedding_model_client import StellaEmbeddings
 from PhD_RAG.src.config import MODEL_CONFIG, MILVUS_CONFIG
@@ -5,9 +6,9 @@ from langchain_core.documents import Document
 
 
 def setup_vectorstore(documents: list[Document], uuids: list[str]) -> Milvus:
-    embeddings = StellaEmbeddings(MODEL_CONFIG['name'])
+    embeddings: StellaEmbeddings = StellaEmbeddings(MODEL_CONFIG['name'])
 
-    vector_store = Milvus(
+    vector_store: Milvus = Milvus(
         embedding_function=embeddings,
         connection_args={"uri": MILVUS_CONFIG['uri']},
         collection_name=MILVUS_CONFIG['collection_name'],
@@ -18,10 +19,24 @@ def setup_vectorstore(documents: list[Document], uuids: list[str]) -> Milvus:
         }
     )
 
-    # Add error handling for document insertion
     try:
         vector_store.add_documents(documents, ids=uuids)
     except Exception as e:
         raise Exception(f"Failed to add documents to vector store: {str(e)}")
 
     return vector_store
+
+
+def chunk_documents(doc_path: str) -> list[Document]:
+    # read file
+    with open(doc_path, encoding='utf-8') as f:
+        markdown_file: str = f.read()
+
+    # define split points
+    headers_to_split_on: list[tuple[str, str]] = [("#", "Header_1")]
+
+    # split text
+    splitter: MarkdownHeaderTextSplitter = MarkdownHeaderTextSplitter(headers_to_split_on)
+    splits: list[Document] = splitter.split_text(markdown_file)
+
+    return splits
